Move the GB holidays in a dedicated file
